refactor: move wa publisher to its own package
